end of jwt authentication, frontend next

+++ backend/controllers/goalsControllers.js
@@ -1,10 +1,11 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -17,7 +18,7 @@ const createGoal = asyncHandler(async (req, res) => {
     throw new Error("Please provide a text");
   }
 
-  const goal = await Goal.create({ text: req.body.text });
+  const goal = await Goal.create({ text: req.body.text, user: req.user.id });
   res.status(200).json(goal);
 });
 
@@ -30,6 +31,20 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+
+  const user = await User.findById(req.user.id);
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User Not Found");
+  }
+
+  // Make sure the logged in user matches the goal user
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -40,6 +55,20 @@ const updateGoal = asyncHandler(async (req, res) => {
 // @route GET /api/goals/:id
 // @access Private
 const deleteGoal = asyncHandler(async (req, res) => {
+  const user = await User.findById(req.user.id);
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User Not Found");
+  }
+
+  // Make sure the logged in user matches the goal user
+
+  const goal = await Goal.findById(req.params.id);
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
   await Goal.findByIdAndDelete(req.params.id);
   res.status(200).json({ id: req.params.id });
 });
